backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
google-generativeai
python-dotenv

